ftpupload migrated to ftpmanager and  static leaks fixed
